add a bunch of blocking comments
