+++ src/index.ts
@@ -1,4 +1,5 @@
-export type {Task, TaskContext, Args, ArgsSchema, ArgSchema, ArgsProperties} from './task/task.js';
+export type {Task, TaskContext} from './task/task.js';
+export type {Args, ArgsSchema, ArgSchema, ArgsProperties} from './utils/args.js';
 export type {Gen, GenContext} from './gen/gen.js';
 
 // export all of the main config helpers and types
