+++ src/typecheck.task.ts
@@ -10,11 +10,13 @@ export const task: Task<TypecheckTaskArgs> = {
 	run: async ({fs, args}): Promise<void> => {
 		const {tsconfig} = args;
 
+		// TODO BLOCK
 		const tscTypecheckResult = await spawn('npx', ['tsc', '--noEmit']);
 		if (!tscTypecheckResult.ok) {
 			throw new TaskError(`Failed to typecheck. ${printSpawnResult(tscTypecheckResult)}`);
 		}
 		if (await fs.exists('node_modules/.bin/svelte-check')) {
+			// TODO BLOCK
 			const svelteCheckResult = await spawn('npx', ['svelte-check', '--tsconfig', tsconfig]);
 			if (!svelteCheckResult.ok) {
 				throw new TaskError(`Failed to typecheck Svelte. ${printSpawnResult(svelteCheckResult)}`);
